refactor: 요청/응답 DTO를 dto.request 및 dto.response 패키지로 분리하고, CommentCreateResponse를 추가하여 댓글 생성 API 응답에 적용 (#98)
