+++ backend/src/main/java/com/tamnara/backend/comment/dto/response/CommentCreateResponse.java
@@ -1,0 +1,10 @@
+package com.tamnara.backend.comment.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CommentCreateResponse {
+    Long commentId;
+}
